Fix MacOS host name add/remove and IP loopback address
The Sed commands provided in removeOldEntriesFromHostfile() and addNewHostname() were causing errors in MacOS on server start. In the MacOS version of Sed, the -i option to edit files in-place requires an extension (which can be blank) for saving backups.  Both functions were updated to include empty single quotes after the -i option.

The Sed command in removeOldEntriesFromHostfile() was leaving empty new lines after removing host names. The regex was updated to look for a pattern and remove the matching line.

In getNewIP() the cfexecute was not updating the loopback address as intended.  This was updated to use the exiting sudo().